Overhaul the project structure and add Dockerfile to build container image without operator-sdk
