+++ api/apis.go
@@ -1,4 +1,4 @@
-package apis
+package api
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
